refactor: initial typescript refactorings
