+++ .gitignore
@@ -551,6 +551,7 @@ google-services.json
 GoogleService-Info.plist
 
 # generated files
+src/version.ts
 RNFBVersion.m
 ReactNativeFirebaseVersion.java
