--- lib/version.js
@@ -1,2 +1,0 @@
-// generated by genversion
-module.exports = '1.0.0';
