+++ src/AdEventType.ts
