+++ src/AdsConsentStatus.ts
