+++ src/MaxAdContentRating.ts
