+++ src/common/Base64.js
