+++ src/common/ReferenceBase.js
