+++ src/common/path.js
