+++ src/common/promise.js
