+++ src/common/serialize.js
