+++ src/internal/GoogleAdsNativeEventEmitter.js
