+++ src/internal/Module.js
