+++ src/internal/NativeError.js
