+++ src/internal/SharedEventEmitter.js
