+++ src/internal/registry/nativeModule.js
