+++ src/types/TestIds.ts
@@ -1,0 +1,8 @@
+/**
+ * Ad Unit IDs used for testing purposes. These should not be used in production apps.
+ */
+export interface TestIds {
+  BANNER: string;
+  INTERSTITIAL: string;
+  REWARDED: string;
+}
