+++ src/validateAdRequestConfiguration.js
