+++ src/validateAdShowOptions.js
