+++ tsconfig.json
@@ -1,13 +1,15 @@
 {
-  "include": ["lib/*.d.ts"],
+  "include": ["src"],
   "compilerOptions": {
     "noEmit": true,
-    "target": "es5",
+    "target": "esnext",
     "module": "commonjs",
     "declaration": true,
     "importHelpers": true,
     "jsx": "react",
     "sourceMap": true,
+    "outDir": "./dist",
+    "rootDir": "./src",
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
@@ -17,7 +19,7 @@
     "skipLibCheck": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
-    "lib": ["es2015", "es2016", "esnext"]
+    "lib": ["esnext"]
   },
   "exclude": ["node_modules", "**/*.spec.ts"]
 }
